+++ Client/BlockProcessing/Game/Features/Terrain/TerrainManager.h
@@ -14,7 +14,7 @@ class TerrainManager {
 public:
     void init(CubeManager* cubeManager, BlockManager* blockManager, ChunkManager* chunkManager, WorldManager* worldManager, int seed, FastNoise::NoiseType noiseType, float frequency, int octaves);
     void generate(const Coord& playerChunkCoord);
-    void render(Mat4f& projectionView, Mat4f& view);
+    void render(Mat4d& projectionView, Mat4d& view);
     void setProjection(Mat4f& projection);
     void setLightPosition(double x, double y, double z);
     ~TerrainManager();
